move auth out part 1
