+++ apps/ui-web/src/App.tsx
@@ -6,25 +6,27 @@ import {
   createRoutesFromElements,
   Route,
   RouterProvider,
+  useNavigate,
 } from "react-router-dom";
+
+import { Link as ReactLink } from "react-router-dom";
+import Link from "@mui/material/Link";
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 
-import {
-  Dashboard,
-  Repo,
-  Test,
-  Profile,
-  Docs,
-  SignIn,
-  SignUp,
-} from "@codepod/ui";
+import { Dashboard, Repo, Test, Docs } from "@codepod/ui";
+
+import { Profile } from "./pages/profile";
+import { SignIn } from "./pages/login";
+import { SignUp } from "./pages/signup";
 
 import { Header, Footer } from "@codepod/ui";
-import { AuthProvider } from "@codepod/ui";
+import { AuthProvider, useAuth } from "./lib/auth";
+import { useMe } from "@codepod/ui";
 
 import Box from "@mui/material/Box";
 import { SnackbarProvider } from "notistack";
+import { Button } from "@mui/material";
 
 const yjsWsUrl = import.meta.env.VITE_APP_YJS_WS_URL;
 const apiUrl = import.meta.env.VITE_APP_API_URL;
@@ -57,9 +59,42 @@ const NormalLayout: React.FC<NormalLayoutProps> = ({
   currentPage,
   children,
 }) => {
+  const { isSignedIn, signOut } = useAuth();
+  let navigate = useNavigate();
+  const { me } = useMe();
   return (
     <Box>
-      <Header currentPage={currentPage} />
+      <Header currentPage={currentPage}>
+        {isSignedIn() ? (
+          <Box
+            sx={{
+              display: "flex",
+              flexDirection: "row",
+              alignItems: "center",
+            }}
+          >
+            <Box sx={{ mr: 2 }}>
+              <Link component={ReactLink} to="/profile" underline="none">
+                {me?.firstname}
+              </Link>
+            </Box>
+            <Button
+              onClick={() => {
+                signOut();
+                navigate("/login");
+              }}
+            >
+              Logout
+            </Button>
+          </Box>
+        ) : (
+          <Box display="block">
+            <Link to="/login" component={ReactLink} underline="none">
+              Login
+            </Link>
+          </Box>
+        )}
+      </Header>
       <Box pt="50px">{children}</Box>
       {/* <Footer /> */}
     </Box>
