+++ apps/ui/src/index.tsx
@@ -1,9 +1,6 @@
 export * from "./pages/test";
 export * from "./pages/index";
-export * from "./pages/login";
-export * from "./pages/signup";
 export * from "./pages/dashboard";
-export * from "./pages/profile";
 export * from "./pages/repo";
 export * from "./pages/docs";
